// src/apis/models/FireBytes_Unified_Wave_Zones_Dtos_ZoneEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_Zones_Dtos_ZoneEditDto = {
    id?: string | null;
    tenantId?: string | null;
    zoneName?: string | null;
    description?: string | null;
    arm?: boolean | null;
    deviceList?: Array<string> | null;
};