// src/apis/models/FireBytes_Unified_Shared_NameVerifyInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Shared_NameVerifyInput = {
    name?: string | null;
    id?: string | null;
    deviceId?: string | null;
    deviceTypeId?: string | null;
    deviceTypeFunctionId?: string | null;
    code?: string | null;
    channel?: string | null;
};